alteracoes no banco de dados, criacao de relacionamentos e login funcionando, controllers com algumas funcionalidades aplicadas e outras somente em comentarios

+++ database/migrations/2022_03_03_144412_create_company_plan_table.php
@@ -16,8 +16,8 @@ class CreateCompanyPlanTable extends Migration
         Schema::create('company_plan', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('price');
-            $table->float('quantity');
+            $table->decimal('price');
+            $table->integer('quantity');
             $table->timestamps();
         });
     }
